introduced config.load(), push50 now looks for files:
