+++ tests/push50_tests.py
@@ -10,161 +10,8 @@ import tempfile
 import logging
 import termcolor
 import subprocess
-
 import push50
-
-class TestGit(unittest.TestCase):
-    def setUp(self):
-        self.info_output = []
-        self.debug_output = []
-
-        self.old_info = push50.logger.info
-        self.old_debug = logging.debug
-
-        push50.logger.info = lambda msg : self.info_output.append(msg)
-        push50.logger.debug = lambda msg : self.debug_output.append(msg)
-
-    def tearDown(self):
-        push50.logger.info = self.old_info
-        push50.logger.debug = self.old_debug
-
-    def test_no_args(self):
-        self.assertEqual(push50.Git()("foo"), "git foo")
-        self.assertEqual(self.info_output, [termcolor.colored("git foo", attrs=["bold"])])
-        self.assertTrue(self.debug_output, ["git foo"])
-
-    def test_arg(self):
-        self.assertEqual(push50.Git().set("bar")("foo"), "git bar foo")
-        self.assertEqual(self.info_output, [termcolor.colored("git bar foo", attrs=["bold"])])
-        self.assertTrue(self.debug_output, ["git bar foo"])
-
-    def test_args(self):
-        self.assertEqual(push50.Git("bar").set("baz")("foo"), "git bar baz foo")
-        self.assertEqual(self.info_output, [termcolor.colored("git bar baz foo", attrs=["bold"])])
-        self.assertTrue(self.debug_output, ["git bar baz foo"])
-
-    def test_special_args_not_set(self):
-        try:
-            push50.Git.work_tree = "bar"
-            push50.Git.git_dir = "baz"
-            push50.Git.cache = "qux"
-
-            self.assertEqual(push50.Git()("foo"), "git foo")
-            self.assertEqual(self.info_output, [termcolor.colored("git foo", attrs=["bold"])])
-            self.assertTrue(self.debug_output, ["git foo"])
-        finally:
-            push50.Git.work_tree = ""
-            push50.Git.git_dir = ""
-            push50.Git.cache = ""
-
-    def test_special_args(self):
-        try:
-            push50.Git.working_area = "bar"
-            push50.Git.cache = "baz"
-
-            git = push50.Git(push50.Git.working_area, push50.Git.cache)
-            self.assertEqual(git("foo"), "git bar baz foo")
-            self.assertEqual(self.info_output, [termcolor.colored("git foo", attrs=["bold"])])
-            self.assertTrue(self.debug_output, ["git bar baz foo"])
-        finally:
-            push50.Git.working_area = ""
-            push50.Git.cache = ""
-
-class TestSlug(unittest.TestCase):
-    def test_wrong_format(self):
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("/cs50/problems2/foo/bar")
-
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("cs50/problems2/foo/bar/")
-
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("/cs50/problems2/foo/bar/")
-
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("cs50/problems2")
-
-    def test_online(self):
-        slug = push50.Slug("cs50/problems2/foo/bar")
-        self.assertEqual(slug.slug, "cs50/problems2/foo/bar")
-        self.assertEqual(slug.org, "cs50")
-        self.assertEqual(slug.repo, "problems2")
-        self.assertEqual(slug.branch, "foo")
-        self.assertEqual(slug.problem, pathlib.Path("bar"))
-
-    def test_wrong_slug_online(self):
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("cs50/does/not/exist")
-
-    def test_offline(self):
-        try:
-            old_local_path = push50.LOCAL_PATH
-            old_wd = os.getcwd()
-
-            push50.LOCAL_PATH = tempfile.TemporaryDirectory().name
-            path = pathlib.Path(push50.LOCAL_PATH) / "foo" / "bar" / "baz"
-            os.makedirs(path)
-
-            os.chdir(pathlib.Path(push50.LOCAL_PATH) / "foo" / "bar")
-            subprocess.check_output(["git", "init"])
-
-            os.chdir(path)
-
-            with open(".cs50.yaml", "w") as f:
-                pass
-            subprocess.check_output(["git", "add", ".cs50.yaml"])
-            out = subprocess.check_output(["git", "commit", "-m", "qux"])
-
-            slug = push50.Slug("foo/bar/master/baz", offline=True)
-            self.assertEqual(slug.slug, "foo/bar/master/baz")
-            self.assertEqual(slug.org, "foo")
-            self.assertEqual(slug.repo, "bar")
-            self.assertEqual(slug.branch, "master")
-            self.assertEqual(slug.problem, pathlib.Path("baz"))
-        finally:
-            push50.LOCAL_PATH = old_local_path
-            os.chdir(old_wd)
-
-    def test_wrong_slug_offline(self):
-        with self.assertRaises(push50.InvalidSlugError):
-            push50.Slug("cs50/does/not/exist", offline=True)
-
-class TestProgressBar(unittest.TestCase):
-    def test_progress(self):
-        f = io.StringIO()
-        with contextlib.redirect_stdout(f):
-            with push50.ProgressBar("foo"):
-                pass
-        self.assertTrue("foo..." in f.getvalue())
-
-    def test_progress_moving(self):
-        f = io.StringIO()
-        with contextlib.redirect_stdout(f):
-            try:
-                old_ticks_per_second = push50.ProgressBar.TICKS_PER_SECOND
-                push50.ProgressBar.TICKS_PER_SECOND = 100
-                with push50.ProgressBar("foo"):
-                    time.sleep(.5)
-            finally:
-                push50.ProgressBar.TICKS_PER_SECOND = old_ticks_per_second
-
-        self.assertTrue("foo...." in f.getvalue())
-
-    def test_disabled(self):
-        f = io.StringIO()
-        with contextlib.redirect_stdout(f):
-            try:
-                old_disabled = push50.ProgressBar.DISABLED
-                push50.ProgressBar.DISABLED = True
-                old_ticks_per_second = push50.ProgressBar.TICKS_PER_SECOND
-                push50.ProgressBar.TICKS_PER_SECOND = 100
-                with push50.ProgressBar("foo"):
-                    time.sleep(.5)
-            finally:
-                push50.ProgressBar.DISABLED = old_disabled
-                push50.ProgressBar.TICKS_PER_SECOND = old_ticks_per_second
-
-        self.assertEqual("foo...\n", f.getvalue())
+import push50.config
 
 class TestConnect(unittest.TestCase):
     def setUp(self):
@@ -219,9 +66,12 @@ class TestFiles(unittest.TestCase):
         os.chdir(self._wd)
 
     def test_exclude_only_one(self):
-        config = {
-            "exclude" : ["foo.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude foo.py\n"
+
+        config = push50.config.load(content, "check50")
 
         open("foo.py", "w").close()
         open("bar.py", "w").close()
@@ -231,9 +81,12 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"foo.py"})
 
     def test_exclude_all(self):
-        config = {
-            "exclude" : ["*"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*\"\n"
+
+        config = push50.config.load(content, "check50")
 
         included, excluded = push50.files(config)
         self.assertEqual(included, set())
@@ -246,9 +99,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"foo.py"})
 
     def test_include_only_one(self):
-        config = {
-            "exclude" : ["*", "!foo.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*\"\n" \
+            "    - !include foo.py\n"
+
+        config = push50.config.load(content, "check50")
 
         open("foo.py", "w").close()
         open("bar.py", "w").close()
@@ -267,18 +124,24 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(included), {"foo.py", "bar.c"})
         self.assertEqual(set(excluded), set())
 
-        config = {
-            "exclude" : ["!*"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !include \"*\"\n"
+
+        config = push50.config.load(content, "check50")
 
         included, excluded = push50.files(config)
         self.assertEqual(set(included), {"foo.py", "bar.c"})
         self.assertEqual(set(excluded), set())
 
     def test_required(self):
-        config = {
-            "required" : ["foo.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !require foo.py\n"
+
+        config = push50.config.load(content, "check50")
 
         open("foo.py", "w").close()
 
@@ -293,10 +156,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), set())
 
     def test_required_overwrite_exclude(self):
-        config = {
-            "exclude" : ["*"],
-            "required" : ["foo.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*\"\n" \
+            "    - !require foo.py\n"
+
+        config = push50.config.load(content, "check50")
 
         open("foo.py", "w").close()
 
@@ -311,9 +177,12 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"bar.c"})
 
     def test_always_exclude(self):
-        config = {
-            "exclude" : ["!foo.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !include foo.py\n"
+
+        config = push50.config.load(content, "check50")
 
         open("foo.py", "w").close()
 
@@ -322,9 +191,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), set())
 
     def test_exclude_folder_include_file(self):
-        config = {
-            "exclude" : ["foo", "!foo/bar"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude foo\n" \
+            "    - !include foo/bar\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         open("foo/bar", "w").close()
@@ -334,9 +207,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), set())
 
     def test_include_file_exclude_folder(self):
-        config = {
-            "exclude" : ["!foo/bar.py", "foo"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !include foo/bar.py\n" \
+            "    - !exclude foo\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         open("foo/bar.py", "w").close()
@@ -346,9 +223,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"foo/bar.py"})
 
     def test_exclude_extension_include_folder(self):
-        config = {
-            "exclude" : ["*.py", "!foo"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*.py\"\n" \
+            "    - !include foo\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         open("foo/bar.py", "w").close()
@@ -358,9 +239,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), set())
 
     def test_exclude_extension_include_everything_from_folder(self):
-        config = {
-            "exclude" : ["*.py", "!foo/*"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*.py\"\n" \
+            "    - !include \"foo/*\"\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         open("foo/bar.py", "w").close()
@@ -370,9 +255,13 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), set())
 
     def test_exclude_everything_include_folder(self):
-        config = {
-            "exclude" : ["*", "!foo"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*\"\n" \
+            "    - !include foo\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         open("foo/bar.py", "w").close()
@@ -386,26 +275,35 @@ class TestFiles(unittest.TestCase):
         open("foo/bar.py", "w").close()
         open("qux.py", "w").close()
 
-        config = {
-            "exclude" : ["*.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*.py\"\n"
+
+        config = push50.config.load(content, "check50")
 
         included, excluded = push50.files(config)
         self.assertEqual(set(included), set())
         self.assertEqual(set(excluded), {"qux.py", "foo/bar.py"})
 
-        config = {
-            "exclude" : ["./*.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"./*.py\"\n"
+
+        config = push50.config.load(content, "check50")
 
         included, excluded = push50.files(config)
         self.assertEqual(set(included), {"foo/bar.py"})
         self.assertEqual(set(excluded), {"qux.py"})
 
     def test_implicit_recursive_with_slash(self):
-        config = {
-            "exclude" : ["*/*.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"*/*.py\"\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         os.mkdir("foo/bar")
@@ -417,9 +315,12 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"foo/qux.py"})
 
     def test_explicit_recursive(self):
-        config = {
-            "exclude" : ["foo/**/*.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !exclude \"foo/**/*.py\"\n"
+
+        config = push50.config.load(content, "check50")
 
         os.mkdir("foo")
         os.mkdir("foo/bar")
@@ -432,18 +333,25 @@ class TestFiles(unittest.TestCase):
         self.assertEqual(set(excluded), {"foo/bar/baz/qux.py"})
 
     def test_requires_no_exclude(self):
-        config = {
-            "required": ["does_not_exist.py"]
-        }
+        content = \
+            "check50:\n" \
+            "  files:\n" \
+            "    - !require does_not_exist.py\n"
+
+        config = push50.config.load(content, "check50")
 
         with self.assertRaises(push50.MissingFilesError):
             push50.files(config)
 
     def test_invalid_utf8_filename(self):
-        open(b"\xc3\x28", "w").close()
-        included, excluded = push50.files({})
-        self.assertEqual(included, set())
-        self.assertEqual(excluded, {"?("})
+        try:
+            open(b"\xc3\x28", "w").close()
+        except OSError:
+            self.skipTest("can't create invalid utf8 filename")
+        else:
+            included, excluded = push50.files({})
+            self.assertEqual(included, set())
+            self.assertEqual(excluded, {"?("})
 
 class TestLocal(unittest.TestCase):
     def setUp(self):
